Main script for menu

+++ src/modify.py
@@ -8,38 +8,8 @@ import json
 import math
 import time
 from treeGen import ArtificialTree
-from utils import getTimeInMs
+from utils import getTimeInMs, parse_dist_params, parse_edit_params
 SAVE_TO_DIR = "../data/"
-
-def parse_edit_params(edit_mode="default"):
-    
-    # Load JSON from file
-    with open("dists/params.json", 'r') as f:
-        params = json.load(f)
-
-    # Degree
-    modify_dist = {k: v for k, v in params["modify"][edit_mode].items()}
-    
-    return modify_dist
-
-def parse_dist_params(degree_mode="low_degree", filetype_mode="default", file_ext_mode="default", permissions_mode="default", size_mode="default"):
-    
-    # Load JSON from file
-    with open("dists/params.json", 'r') as f:
-        params = json.load(f)
-
-    # Degree
-    degree_distr = {int(k): v for k, v in params["degree"][degree_mode].items()}
-    # File type
-    fileType_distr = {k: v for k, v in params["type"][filetype_mode].items()}
-    # Extensions
-    fileExt_distr = {k: v for k, v in params["filetypes"][file_ext_mode].items()}
-    # Permissions
-    permissions_distr = {int(k, 8): v for k, v in params["permissions"][permissions_mode].items()}
-    # Sizes
-    size_distr = {k: v for k, v in params["size"][size_mode].items()}
-
-    return degree_distr, fileType_distr, fileExt_distr, permissions_distr, size_distr
 
 class ModifyTree:
     def __init__(self, originalRoot, newRoot, maxNew, iterations):
@@ -47,6 +17,8 @@ class ModifyTree:
         self.originalRoot = Path(f"{SAVE_TO_DIR}{originalRoot}")
         # Max new files to be added to the new tree
         self.maxNew = maxNew
+        # Track files
+        self.filesMade = 0
         # User mode
         self.userMode = True
         # Dict of edit probabilities
@@ -56,6 +28,7 @@ class ModifyTree:
             self.editProbs = parse_edit_params()
         # Level of user home dirs
         self.userHome = self.editProbs["root"]
+        print(self.userHome)
         # Iterations, num of trees to generate
         self.iterations = iterations
         # List to easily maintain each iteration of the modification
@@ -100,6 +73,8 @@ class ModifyTree:
                 self.modify_recurse(newRoot, 0, 0)
             endModifyTime = time.perf_counter()
             print(f"Modified tree '{str(oldRoot)[8:]}' in {getTimeInMs(startModifyTime, endModifyTime)} ms.")
+            print(f"Made {self.filesMade} new files this iteration.\n")
+            self.filesMade = 0
         
     # Recursively edit a tree
     def modify_recurse(self, root, files_made, current_depth):
@@ -115,7 +90,10 @@ class ModifyTree:
                         shutil.rmtree(entry)
                         continue
                     # If not deleted, recurse into the subdir
-                    files_made += self.modify_recurse(entry, files_made, current_depth+1)
+                    
+                    newfiles = self.modify_recurse(entry, files_made, current_depth+1)
+                    files_made += newfiles
+                    self.filesMade += newfiles
                 
                 # Files
                 elif entry.is_file():
@@ -155,11 +133,15 @@ class ModifyTree:
                         fileText = "".join(ran.choices(string.ascii_letters + string.digits, k=size))
                         f.write(fileText)
                     files_made += 1
+                    self.filesMade += 1
 
         if files_made < self.maxNew:
             # Create folders
             if ran.random() < self.editProbs["createFolder"]:
-                files_made += tempTree.gen_tree_atlevel(treeCreateDepth, root, max_files=max(0, self.maxNew - files_made))
+                
+                new_files = tempTree.gen_tree_atlevel(treeCreateDepth, root, max_files=max(0, self.maxNew - files_made))
+                files_made += new_files
+                self.filesMade += files_made
         
         return files_made
 
@@ -170,13 +152,16 @@ class ModifyTree:
         for entry in root.iterdir():
             if files_made < self.maxNew:
                 # Folders
-                if entry.is_dir() and self.userHome <= current_depth:
+                if entry.is_dir():
                     # Delete based on probability
-                    if ran.random() < self.editProbs["deleteFolder"]:
+                    if ran.random() < self.editProbs["deleteFolder"]  and self.userHome <= current_depth:
                         shutil.rmtree(entry)
                         continue
                     # If not deleted, recurse into the subdir
-                    files_made += self.modify_recurse(entry, files_made, current_depth+1)
+                    
+                    newfiles = self.modify_recurse_user(entry, files_made, current_depth+1)
+                    files_made += newfiles
+                    self.filesMade += newfiles
                 
                 # Files
                 elif entry.is_file() and self.userHome <= current_depth:
@@ -221,7 +206,10 @@ class ModifyTree:
         if files_made < self.maxNew and self.userHome <= current_depth:
             # Create folders
             if ran.random() < self.editProbs["createFolder"]:
-                files_made += tempTree.gen_tree_atlevel(treeCreateDepth, root, max_files=max(0, self.maxNew - files_made))
+                
+                newfiles = tempTree.gen_tree_atlevel(treeCreateDepth, root, max_files=max(0, self.maxNew - files_made))
+                files_made += newfiles
+                self.filesMade += newfiles
         
         return files_made
